wip: main cleanup done
