+++ elixir/web-ng/assets/js/lib/god_view/lifecycle_bootstrap_channel_socket_methods.js
@@ -1,0 +1,11 @@
+import {Socket} from "phoenix"
+
+export const godViewLifecycleBootstrapChannelSocketMethods = {
+  ensureGodViewSocket() {
+    if (!window.godViewSocket) {
+      window.godViewSocket = new Socket("/socket", {params: {_csrf_token: this.csrfToken}})
+      window.godViewSocket.connect()
+    }
+    return window.godViewSocket
+  },
+}
