+++ elixir/web-ng/assets/js/lib/god_view/lifecycle_bootstrap_cleanup_dom_methods.js
@@ -1,0 +1,10 @@
+export const godViewLifecycleBootstrapCleanupDomMethods = {
+  cleanupLifecycleDomListeners() {
+    window.removeEventListener("resize", this.resizeCanvas)
+    if (this.canvas) this.canvas.removeEventListener("wheel", this.handleWheelZoom)
+    if (this.canvas) this.canvas.removeEventListener("pointerdown", this.handlePanStart)
+    window.removeEventListener("pointermove", this.handlePanMove)
+    window.removeEventListener("pointerup", this.handlePanEnd)
+    window.removeEventListener("pointercancel", this.handlePanEnd)
+  },
+}
