+++ elixir/web-ng/assets/js/lib/god_view/lifecycle_bootstrap_cleanup_methods.js
@@ -1,24 +1,16 @@
-export const godViewLifecycleBootstrapCleanupMethods = {
+import {godViewLifecycleBootstrapCleanupDomMethods} from "./lifecycle_bootstrap_cleanup_dom_methods"
+import {godViewLifecycleBootstrapCleanupRuntimeMethods} from "./lifecycle_bootstrap_cleanup_runtime_methods"
+
+const godViewLifecycleBootstrapCleanupCoreMethods = {
   cleanupLifecycle() {
-    window.removeEventListener("resize", this.resizeCanvas)
-    if (this.canvas) this.canvas.removeEventListener("wheel", this.handleWheelZoom)
-    if (this.canvas) this.canvas.removeEventListener("pointerdown", this.handlePanStart)
-    window.removeEventListener("pointermove", this.handlePanMove)
-    window.removeEventListener("pointerup", this.handlePanEnd)
-    window.removeEventListener("pointercancel", this.handlePanEnd)
-    this.stopAnimationLoop()
-    this.stopPolling()
-    if (this.channel) {
-      this.channel.leave()
-      this.channel = null
-    }
-    if (this.pendingAnimationFrame) {
-      cancelAnimationFrame(this.pendingAnimationFrame)
-      this.pendingAnimationFrame = null
-    }
-    if (this.deck) {
-      this.deck.finalize()
-      this.deck = null
-    }
+    this.cleanupLifecycleDomListeners()
+    this.cleanupLifecycleRuntime()
   },
 }
+
+export const godViewLifecycleBootstrapCleanupMethods = Object.assign(
+  {},
+  godViewLifecycleBootstrapCleanupCoreMethods,
+  godViewLifecycleBootstrapCleanupDomMethods,
+  godViewLifecycleBootstrapCleanupRuntimeMethods,
+)
